requirements.txt
easyocr>=1.7
pymupdf>=1.24
chromadb>=0.5
google-genai>=1.0
python-dotenv>=1.0
rich>=13.0
